BUG: Add numpy headers and other settings to setup
gcc was failing because it wasn't able to find the location of the numpy
headers.

The module could not be imported because it was not declared as a
package in setup.py and it needed to have the diptest/dip_crit.txt file
as part of the package date for the module to work.